benchmark db client
